% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iMAAPssmary0.2.R
\name{iMAAPssmry}
\alias{iMAAPssmry}
\title{Summary the output of iMAAPs}
\usage{
iMAAPssmry(ad_inp, rawld_inp, num_chr, denoise_cut, method = "median")
}
\arguments{
\item{output}{files of iMAAPs, the number of chromosomes, denoise cutoff, method}
}
\value{
Summary of iMAAPs output
}
\description{
This function accepts the output files (.rawld and .ad)
of iMAAPs and summarizes the data.
Required arguments:
ad_inp is a matrix of the output fitting coeffients
rawld_inp is a matrix of the output Weighted LD decays
denoise_cut is a number between 0 and 1 define the noise cut threshold
num_chr is a number of the chromosome used for Jacknife process
method is a function used to estimate the average of the admixture signal, it has two choices: median and mean
}
